--- a/OpenClaw/Engine/Actor/Components/AIComponents/ProjectileAIComponent.cpp
+++ b/OpenClaw/Engine/Actor/Components/AIComponents/ProjectileAIComponent.cpp
@@ -35,7 +35,7 @@ ProjectileAIComponent::ProjectileAIComponent()
 
 ProjectileAIComponent::~ProjectileAIComponent()
 {
-    for (auto pSparkle : m_PowerupSparkles)
+    for (const auto &pSparkle : m_PowerupSparkles)
     {
         shared_ptr<EventData_Destroy_Actor> pEvent(new EventData_Destroy_Actor(pSparkle->GetGUID()));
         IEventMgr::Get()->VQueueEvent(pEvent);
@@ -80,6 +80,7 @@ void ProjectileAIComponent::VPostInit()
         m_pPhysics->VSetLinearSpeed(m_pOwner->GetGUID(), m_ProjectileSpeed);
     }
 
+    m_PowerupSparkles.reserve(m_NumSparkles);
     for (int sparkleIdx = 0; sparkleIdx < m_NumSparkles; sparkleIdx++)
     {
         StrongActorPtr pPowerupSparkle = ActorTemplates::CreatePowerupSparkleActor(50);